Splitting rocsparse_handle.hpp
