+++ library/src/include/rocsparse_mat_descr.hpp
@@ -1,0 +1,50 @@
+/*! \file */
+/* ************************************************************************
+ * Copyright (C) 2025 Advanced Micro Devices, Inc. All rights Reserved.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ *
+ * ************************************************************************ */
+
+#pragma once
+
+#include "rocsparse-types.h"
+
+/********************************************************************************
+ * \brief rocsparse_mat_descr is a structure holding the rocsparse matrix
+ * descriptor. It must be initialized using rocsparse_create_mat_descr()
+ * and the retured handle must be passed to all subsequent library function
+ * calls that involve the matrix.
+ * It should be destroyed at the end using rocsparse_destroy_mat_descr().
+ *******************************************************************************/
+struct _rocsparse_mat_descr
+{
+    // matrix type
+    rocsparse_matrix_type type = rocsparse_matrix_type_general;
+    // fill mode
+    rocsparse_fill_mode fill_mode = rocsparse_fill_mode_lower;
+    // diagonal type
+    rocsparse_diag_type diag_type = rocsparse_diag_type_non_unit;
+    // index base
+    rocsparse_index_base base = rocsparse_index_base_zero;
+    // storage mode
+    rocsparse_storage_mode storage_mode = rocsparse_storage_mode_sorted;
+    // maximum nnz per row
+    int64_t max_nnz_per_row = 0;
+};
